# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Billing API",
        "description": "API для управления балансами пользователей и их транзакциями",
        "version": "1.0"
    },
    "paths": {
        "/api/transaction": {
            "post": {
                "tags": [
                    "Транзакции"
                ],
                "summary": "Создание транзакции (пополнение/списание)",
                "operationId": "c1dfbb6c34574495e19a7937dcd548ca",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "amount",
                                    "description"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100.5
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Пополнение баланса"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Транзакция успешно выполнена"
                    },
                    "400": {
                        "description": "Недостаточно средств"
                    }
                }
            }
        },
        "/api/transactions": {
            "get": {
                "tags": [
                    "Транзакции"
                ],
                "summary": "Получение истории транзакций",
                "operationId": "6581469356772e7a74a0d79897544f1a",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Сортировка по дате",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ],
                            "example": "desc"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Поиск по описанию транзакции",
                        "schema": {
                            "type": "string",
                            "example": "Пополнение"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "История транзакций успешно получена"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Пользователь"
                ],
                "summary": "Регистрация нового пользователя",
                "operationId": "2f25885f3b6f61ca11724ec302d2e439",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "johndoe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "securepassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Пользователь успешно зарегистрирован"
                    },
                    "400": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/user/{id}": {
            "get": {
                "tags": [
                    "Пользователь"
                ],
                "summary": "Получение информации о пользователе",
                "operationId": "d8f696420bea235f4916124c9ba1c507",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о пользователе успешно получена"
                    },
                    "404": {
                        "description": "Пользователь не найден"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Транзакции",
            "description": "Транзакции"
        },
        {
            "name": "Пользователь",
            "description": "Пользователь"
        }
    ]
}